% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgmm.R
\name{fitgmm}
\alias{fitgmm}
\title{gmmfit}
\usage{
fitgmm(data, k = 2, maxiter = 10, method = c("full", "diag"))
}
\description{
gmmfit
}
\examples{
x1 = rmvnorm(50, mean=rep(-1,4))
x2 = rmvnorm(50, mean=rep(+1,4))
xx = rbind(x1, x2)

cl1 = fitgmm(xx, k=1)
cl2 = fitgmm(xx, k=2)
cl3 = fitgmm(xx, k=3)

cls = array(0,c(10,3))
for (i in 1:10){
  cls[i,] = as.vector(fitgmm(xx, k=i, method="diag")$criteria)
  print(paste("iteration ",i," complete..",sep=""))
}

yy = c(rnorm(30,mean=-1),rnorm(30,mean=1))
cls = array(0,c(10,3))
for (i in 1:10){
  cls[i,] = as.vector(fitgmm(yy, k=i, method="diag")$criteria)
  print(paste("iteration ",i," complete..",sep=""))
}

}
