% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.barycenter.R
\name{gauss.barycenter}
\alias{gauss.barycenter}
\title{Barycenter of Gaussian Distributions}
\usage{
gauss.barycenter(x, y = NULL, type = c("wass2"))
}
\description{
Barycenter of Gaussian Distributions
}
