% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.median.R
\name{gauss.median}
\alias{gauss.median}
\title{Geometric Median of Gaussian Distributions}
\usage{
gauss.median(
  x,
  y = NULL,
  type = c("wass2"),
  maxiter = 200,
  eps = 1e-10,
  nthreads = 1
)
}
\description{
Geometric Median of Gaussian Distributions
}
