% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.pdist2.R
\name{gauss.pdist2}
\alias{gauss.pdist2}
\title{Pairwise Distances between two sets of Gaussian distributions}
\usage{
gauss.pdist2(glist1, glist2, method = c("wass2"))
}
\description{
Pairwise Distances between two sets of Gaussian distributions
}
