% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.sample.R
\name{gauss.sample}
\alias{gauss.sample}
\title{Draw random samples from Gaussian distribution via 'wrapgauss' object}
\usage{
gauss.sample(n = 100, obj.gauss)
}
\description{
It's a wrapper for 'rmvnorm' and 'rmvnorm'
}
