% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.eval.R
\name{gmm.eval}
\alias{gmm.eval}
\title{Evaluate density and hard clustering of data with GMM model}
\usage{
gmm.eval(mydata, gmmobj)
}
\description{
Evaluate density and hard clustering of data with GMM model
}
