% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.merge.R
\name{gmm.merge}
\alias{gmm.merge}
\title{Large-scale Finite Gaussian Mixture Model Estimation via Split and Merge}
\usage{
gmm.merge(gmmlist, k = 2, ctype = c("kmeans", "kmedoids"))
}
\description{
Large-scale Finite Gaussian Mixture Model Estimation via Split and Merge
}
