% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpgauss.R
\name{interpgauss}
\alias{interpgauss}
\title{Interpolate Two Gaussian Distributions}
\usage{
interpgauss(gauss1, gauss2, t = 0.5, method = c("wass2"))
}
\description{
method (1-t)N0 + tN1
}
\examples{
## test with 1-dimensional Gaussian distributions
obj1 = wrapgauss1d(mean=1, sd=1)
obj5 = wrapgauss1d(mean=5, sd=5)

## let's interpolate
obj.half  = interpgauss(obj1, obj5, t=0.5)


}
