% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{mvnorm}
\alias{mvnorm}
\alias{dmvnorm}
\alias{rmvnorm}
\title{Multivariate Gaussian Distribution}
\usage{
dmvnorm(x, mean = rep(0, ncol(x)), sigma = diag(length(mean)), log = FALSE)

rmvnorm(
  n,
  mean = rep(0, nrow(sigma)),
  sigma = diag(length(mean)),
  method = c("eigen", "chol")
)
}
\arguments{
\item{x}{a param for toBar and notToBar}

\item{y}{a param just for notToBar}
}
\value{
Hard to have one return section for all functions,
might want to have a manual list here.
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

