% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapgauss.R
\name{wrapgauss}
\alias{wrapgauss}
\alias{wrapgauss1d}
\alias{wrapgaussNd}
\title{Wrap Gaussian distribution into \code{S3} object of class 'wrapgauss'}
\usage{
wrapgauss1d(mean, sd)

wrapgaussNd(mu, sigma)
}
\arguments{
\item{mean}{1d}

\item{sd}{1d}

\item{mu}{nd}

\item{sigma}{nd}
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

