% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapgmm.R
\name{wrapgmm}
\alias{wrapgmm}
\title{Wrap Mixture of Gaussians into \code{S3} object of class 'wrapgmm'}
\usage{
wrapgmm(wglist, weight = rep(1/length(wglist), length(wglist)))
}
\description{
Wrap Mixture of Gaussians into \code{S3} object of class 'wrapgmm'
}
