% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.kmeans.R
\name{gauss.kmeans}
\alias{gauss.kmeans}
\title{K-Means clustering for Gaussian Distributions}
\usage{
gauss.kmeans(glist, k = 2, type = c("wass2"), maxiter = 100)
}
\description{
K-Means clustering for Gaussian Distributions
}
\examples{
## generate three-cluster data with univariate Gaussians
mylist = list()
for (i in 1:10){
   mylist[[i]] = wrapgauss1d(mean=-2, sd=1)
}
for (i in 11:20){
   mylist[[i]] = wrapgauss1d(mean=0, sd=1)
}
for (i in 21:30){
   mylist[[i]] = wrapgauss1d(mean=2, sd=1)
}

## apply clustering with different k values
cl2 <- gauss.kmeans(mylist, k=2)$cluster
cl3 <- gauss.kmeans(mylist, k=3)$cluster
cl4 <- gauss.kmeans(mylist, k=4)$cluster

## compute 2-dimensional embedding for visualization
mds2d <- gauss.mds(mylist, ndim=2)$embed
mdsx <- as.vector(mds2d[,1])
mdsy <- as.vector(mds2d[,2])

## visualize
opar = par(mfrow=c(1,3), pty="s")
plot(mdsx, mdsy, pch=19, col=cl2, main="k=2 means")
plot(mdsx, mdsy, pch=19, col=cl3, main="k=3 means")
plot(mdsx, mdsy, pch=19, col=cl4, main="k=4 means")
par(opar)

}
