% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.pdist.R
\name{gauss.pdist}
\alias{gauss.pdist}
\title{Pairwise Distance between Gaussian distributions}
\usage{
gauss.pdist(x, y = NULL, type = c("wass2"), as.dist = FALSE)
}
\description{
Pairwise Distance between Gaussian distributions
}
\examples{
## test with 1-dimensional Gaussian distributions
list1d = list()
for (i in 1:20){
   val.center  = rnorm(1, sd=0.25)
   list1d[[i]] = wrapgauss1d(mean=val.center, sd=1)
}
for (j in 21:40){
   val.center  = rnorm(1, sd=0.25) + 2
   list1d[[j]] = wrapgauss1d(mean=val.center, sd=1)
}

## compute pairwise distance 
d1wass2 = gauss.pdist(list1d)

## visualize
gcol <- gray((0:64)/64)
opar <- par(pty="s")
image(d1wass2[,40:1], col=gcol, axes=FALSE, main="wass2 pairwise distance")
par(opar)

\dontrun{
## test with 5-dimensional Gaussian distributions
list5d = list()
for (i in 1:20){
   vec.center  = rmvnorm(1, mean=rep(0,5))
   list5d[[i]] = wrapgaussNd(mu=vec.center, sigma=diag(5))
}
mysig = matrix(rnorm(100*5), ncol=5)
mysig = t(mysig)\%*\%mysig
for (j in 21:40){
   vec.center  = rmvnorm(1, mean=rep(0,5)+1.5)
   list5d[[j]] = wrapgaussNd(mu=vec.center, sigma=mysig)
}

## compute pairwise distance
d5wass2 = gauss.pdist(list5d)

## visualize
gcol <- gray((0:64)/64)
opar <- par(pty="s")
image(d5wass2[,40:1], col=gcol, axes=FALSE, main="wass2 pairwise distance")
par(opar)
}

}
